Primitive for binding of caller, and move Kernel#binding to ruby.
