+++ src/main/java/org/truffleruby/core/binding/BindingNodes.java
@@ -14,16 +14,20 @@ import java.util.Set;
 
 import org.truffleruby.Layouts;
 import org.truffleruby.RubyContext;
+import org.truffleruby.builtins.CallerFrameAccess;
 import org.truffleruby.builtins.CoreClass;
 import org.truffleruby.builtins.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.CoreMethodNode;
+import org.truffleruby.builtins.Primitive;
+import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
 import org.truffleruby.builtins.UnaryCoreMethodNode;
 import org.truffleruby.core.array.ArrayHelpers;
 import org.truffleruby.core.cast.NameToJavaStringNodeGen;
 import org.truffleruby.language.RubyGuards;
 import org.truffleruby.language.RubyNode;
 import org.truffleruby.language.Visibility;
+import org.truffleruby.language.arguments.ReadCallerFrameNode;
 import org.truffleruby.language.arguments.RubyArguments;
 import org.truffleruby.language.control.RaiseException;
 import org.truffleruby.language.locals.ReadFrameSlotNode;
@@ -45,6 +49,8 @@ import com.oracle.truffle.api.frame.Frame;
 import com.oracle.truffle.api.frame.FrameDescriptor;
 import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.frame.MaterializedFrame;
+import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.nodes.Node.Child;
 import com.oracle.truffle.api.object.DynamicObject;
 
 @CoreClass("Binding")
@@ -395,4 +401,20 @@ public abstract class BindingNodes {
 
     }
 
+    @Primitive(name = "caller_binding")
+    public abstract static class CallerBindingNode extends PrimitiveArrayArgumentsNode {
+
+        @Child ReadCallerFrameNode callerFrameNode = new ReadCallerFrameNode(CallerFrameAccess.MATERIALIZE);
+
+        public abstract DynamicObject executeBinding(VirtualFrame frame);
+
+        @Specialization
+        public DynamicObject binding(VirtualFrame frame) {
+            // Materialize the caller's frame - false means don't use a slow path to get it - we
+            // want to optimize it
+            final MaterializedFrame callerFrame = callerFrameNode.execute(frame).materialize();
+
+            return BindingNodes.createBinding(getContext(), callerFrame);
+        }
+    }
 }
